Basic signal analysis

SignalAnalysis.energy — Method

energy(s; fs)

Computes total signal energy.

SignalAnalysis.ifrequency — Method

ifrequency(s; fs)

Computes instantaneous frequency of the signal.

SignalAnalysis.meanfrequency — Method

meanfrequency(s; fs, nfft, window)

Computes mean frequency of a signal.

SignalAnalysis.meantime — Method

meantime(s)

Computes mean time of the signal.

SignalAnalysis.padded — Method

padded(s, padding; delay, fill)

Generates a padded view of a signal with optional delay/advance.

SignalAnalysis.rmsbandwidth — Method

rmsbandwidth(s; fs, nfft, window)

Computes RMS bandwidth of a signal.

SignalAnalysis.rmsduration — Method

rmsduration(s)

Computes RMS duration of the signal.

SignalAnalysis.slide — Function

slide(f, s, nframes)
slide(f, s, nframes, overlap, args; showprogress)

Slides a window over a signal, processing each window. If the total number of frames in the signal is not an integral multiple of nframes, the last incomplete block of samples remains unprocessed.

The processing function receives a view on the original signal, and therefore may modify the signal if desired.

Examples:

julia> x = signal(ones(1000), 8kHz);
julia> slide(x, 250) do x1, blknum, firstframe
         println(size(x1), ", ", blknum, ", ", firstframe)
       end
(250,), 1, 1
(250,), 2, 251
(250,), 3, 501
(250,), 4, 751

julia> slide(x, 250) do x1, blknum, firstframe
         x1 .= blknum
       end

julia> x[1], x[251], x[501], x[751]
(1.0, 2.0, 3.0, 4.0)

SignalAnalysis.slide — Method

slide(f, , s, nframes)
slide(f, , s, nframes, overlap, args; showprogress)

Slides a window over a signal, processing each window, and collecting the results. If the total number of frames in the signal is not an integral multiple of nframes, the last incomplete block of samples remains unprocessed.

Examples:

julia> x = signal(ones(1000), 8kHz);
julia> slide(Float32, x, 250) do x1, blknum, firstframe
         sum(x1)*blknum
       end
4-element Array{Float32,1}:
  250.0
  500.0
  750.0
 1000.0

julia> slide(Tuple{Int,Float64}, x, 250) do x1, blknum, firstframe
          (blknum, sum(x1)*blknum)
        end
4-element Array{Tuple{Int64,Float64},1}:
  (1, 250.0)
  (2, 500.0)
  (3, 750.0)
  (4, 1000.0)

SignalAnalysis.toframe — Method

toframe(t, s)

Converts time to signal frame number.

Examples:

julia> x = signal(randn(2000), 8kHz);
julia> toframe(0.2s, x)
1601

julia> toframe([0.2s, 201ms], x)
2-element Array{Int64,1}:
 1601
 1609

julia> toframe(0.2:0.01:0.3, x)
 11-element Array{Int64,1}:
  1601
  1681
  1761
   ⋮