Creating & managing signals

SignalAnalysis.analytic — Method

analytic(s)

Converts a signal to analytic representation.

SignalAnalysis.isanalytic — Method

isanalytic(s)

Checks if a signal is analytic.

SignalAnalysis.samples — Method

samples(s)

Gets the underlying samples in the signal.

SignalAnalysis.signal — Method

signal(x, fs)

Creates a signal with frame rate fs.

SignalAnalysis.signal — Method

signal(x, fs)

Creates a signal with frame rate fs. If the original signal's frame rate is the same as fs, this method simply returns the original signal. Otherwise, it creates a new signal with the specified frame rate and data from the original signal. Do note that this method does not resample the signal.

SignalAnalysis.@rate — Macro

@rate fs expr

Creates a signal from expr with frame rate fs. It provides syntactic sugar on the signal method.

Example:

julia> x = @rate 44.1kHz randn(44100)
44100-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 44100.0Hz

SignalAnalysis.@samerateas — Macro

@samerateas n x expr

Creates a signal from expr with the a frame rate n times that of signal x.

Examples:

julia> x = @rate 44.1kHz randn(44100)
44100-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 44100.0Hz

julia> y = @samerateas 1//2 x x[1:2:end]
22050-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 22050.0Hz

julia> y = @samerateas 2//3 x resample(x, 2//3)
29400-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 29400.0Hz

SignalAnalysis.@samerateas — Macro

@samerateas x expr

Creates a signal from expr with the same frame rate as signal x. This is useful to preserve frame rate metadata across functions that do not return a signal.

Examples:

julia> x = @rate 44.1kHz randn(44100)
44100-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 44100.0Hz

julia> y = filt([1.0,0.5], x)     # frame rate stripped by DSP.filt
44100-element Array{Float64,1}:
   ⋮

julia> y = @samerateas x filt([1.0,0.5], x)
44100-frame, 1-channel SampleBuf{Float64, 1}
1.0s sampled at 44100.0Hz