Time-frequency analysis

SignalAnalysis.tfd — Method

tfd(
    s,
    kernel::Spectrogram;
    onesided,
    fs
) -> SignalAnalysis.TFD{_A} where _A

Computes a spectrogram time-frequency distribution of signal s sampled as sampling rate fs.

Examples:

julia> x = real.(chirp(1kHz, 10kHz, 1s, 44.1kHz));
julia> y = tfd(x, Spectrogram());
julia> plot(y)
julia> y = tfd(x, Spectrogram(nfft=512, noverlap=256, window=hamming));
julia> plot(y)

SignalAnalysis.tfd — Method

tfd(
    s,
    kernel::Wigner;
    onesided,
    fs
) -> SignalAnalysis.TFD{_A} where _A

Computes a Wigner-Ville time-frequency distribution of signal s sampled as sampling rate fs.

Examples:

julia> x = real.(chirp(1kHz, 10kHz, 0.01s, 44.1kHz));
julia> y = tfd(x, Wigner());
julia> plot(y; clim=(0,20))
julia> y = tfd(x, Wigner(nfft=512, smooth=10, method=:CM1980, window=hamming));
julia> plot(y; clim=(0,20))