Plot recipes

Plot recipes are enabled by importing Plots:

using Plots

We provide a plot recipe to display signals:

plot(data::SampleBuf; t0=0.0, downsample=:auto, pooling=:auto, kwargs...)

Optional arguments:

t0=0.0: start time (for labeling only)

downsample=:auto: downsampling factor (integer or :auto)

pooling=:auto: pooling mode (:auto, :min, :max, :mean, :minmax, nothing or function)

If the signal is too long, it is automatically downsampled in a perceptually meaningful way. The downsampling can be controlled using the downsample and pooling keywords.

We also provide convenience plot recipes for common signal processing plots:

SignalAnalysis.psd — Function

psd(data; kwargs...)

Plots the power spectral density of data.

Optional keyword arguments

fs=1.0: derived from the data if a SampleBuf is provided as input

nfft=512: size of FFT window

noverlap=nfft÷2: window overlap size

window=nothing: accepts any window from DSP.jl

xscale=:auto: one of :auto, :identity or :log10

other kwargs are passed on to plot

SignalAnalysis.specgram — Function

specgram(data; kwargs...)

Plots a spectrogram of the data.

Optional keyword arguments

fs=1.0: derived from the data if a SampleBuf is provided as input

nfft=256: size of FFT window

noverlap=nfft÷2: window overlap size

window=nothing: accepts any window from DSP.jl

t0=0.0: start time

downsample=:auto: downsampling factor (integer) for time axis

pooling=:mean: pooling mode (:min, :max, :mean, nothing or function)

other kwargs are passed on to plot

SignalAnalysis.freqresp — Function

freqresp(filter; kwargs...)
freqresp(num; kwargs...)
freqresp(num, den; kwargs...)

Plots frequency response of a digital filter.

Optional keyword arguments

fs=1.0: sampling frequency

nfreq=256: number of frequency points to evaluate filter response at

xscale=:auto: one of :auto, :identity or :log10

other kwargs are passed on to plot